requirements.txt
flask
supabase
opencv-python
numpy
ultralytics
python-dotenv
PyJWT